requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pillow==8.3.1
python-multipart==0.0.5
transformers==4.46.3
torch==2.4.1
requests==2.25.1
python-dotenv==0.19.0
dashscope==1.14.0